+++ tests/test_formatters/test_sh.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python3
 # encoding: utf-8
-from nose import with_setup
-from tests.utils import *
 
+import shlex
 import subprocess
 
+from nose import with_setup
 from parameterized import parameterized
+from tests.utils import *
 
 
 def run_shell_script(shell, sh_path, *args):
@@ -291,3 +292,25 @@ def test_keep_parent_timestamps(shell):
     stat_after = os.stat(dir_path)
 
     assert stat_before.st_mtime == stat_after.st_mtime
+
+
+# regression test for GitHub issue #545
+@parameterized.expand([('',), ('-D',)])
+@with_setup(usual_setup_func, usual_teardown_func)
+def test_skip_hardlinks(tm_opt):
+    dir_a = create_dirs('a')
+    create_file('xxx', 'a/1')
+    create_file('yyy', 'a/2')
+    dir_b = create_dirs('b')
+    create_link('a/2', 'b/2')
+
+    sh_path = os.path.join(TESTDIR_NAME, 'rmlint.sh')
+    run_rmlint(
+        '-S a -o sh:{p} -c sh:hardlink'.format(p=shlex.quote(sh_path)),
+        tm_opt, dir_a, dir_b,
+        use_default_dir=False,
+    )
+
+    counts = pattern_count(sh_path, ["^cp_hardlink +'", "^skip_hardlink +'"])
+    assert counts[0] == 0
+    assert counts[1] == 1
