+++ tests/utils.py
@@ -10,6 +10,7 @@ import json
 import time
 import pprint
 import psutil
+import re
 import shutil
 import shlex
 import struct
@@ -468,3 +469,14 @@ def needs_reflink_fs(test):
         return make_decorator(test)(not_reflink_fs)
     else:
         return test
+
+
+# count the number of line in a file which start with each pattern
+def pattern_count(path, patterns):
+    counts = [0] * len(patterns)
+    with open(path, 'r') as f:
+        for line in f:
+            for i, pattern in enumerate(patterns):
+                if re.match(pattern, line):
+                    counts[i] += 1
+    return counts
